requirements.txt
streamlit==1.44.1
pypdf==5.6.0
transformers==4.52.4
torch==2.6.0
sentencepiece==0.2.0
pandas==2.2.3
